backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
pandas==2.1.4
numpy==1.26.2
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
greenlet==3.0.1
pyee==11.0.1
pinecone-client==3.0.0
sentence-transformers==2.7.0
pinecone-text